+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- **NumericStepper** `disable` prop.
+- **NumericStepper** `readOnly` prop.
 
 ## [9.90.3] - 2019-10-25